refactor(pipeline): extract home classify leaf
Move the home nav-strategy classifier and its href/modal detection
helpers out of HomeResolver into a new leaf module,
HomeStrategyClassify. HomeResolver sat at 148/150 effective lines with
no room for the prefer-DIRECT entry fix that follows next. Behavior is
unchanged: classification runs via the re-exported NAV_STRATEGY barrel.
